refactor: extract area cal
